fix : GET-popular reviews 서비스 코드 완료

+++ src/main/resources/schema.sql
@@ -14,7 +14,7 @@ CREATE TABLE popular_review
     id         BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY NOT NULL,
     review_id  BIGINT                                          NOT NULL,
     rank       SMALLINT                                        NOT NULL,
-    score      SMALLINT                                        NOT NULL,
+    score      DECIMAL(10, 2)                                  NOT NULL,
     period     VARCHAR(10)                                     NOT NULL, -- 'DAILY, MONTLY, YEARLY, ALL'
     created_at timestamp with time zone                        NOT NULL
 );
@@ -99,7 +99,7 @@ CREATE TABLE member
     id         BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY NOT NULL,
     email      VARCHAR(50)                                     NOT NULL UNIQUE,
     nickname   VARCHAR(50)                                     NOT NULL UNIQUE,
-    password   VARCHAR(100)                                     NOT NULL, --'bcrypt'
+    password   VARCHAR(100)                                    NOT NULL, --'bcrypt'
     created_at timestamp with time zone                        NOT NULL,
     updated_at timestamp with time zone                        NULL,
     deleted    BOOLEAN                                         NOT NULL
